App works now with normal List<T>, cloned using ToList()
Removed NList<T>, replaced its use with List<T>, removed unnecessary extensions the ToDoApp1.

To fix the problem with NReact not catching the change when needed, I had to clone the original List using ToList() method, adding item to the created List<T>, and assigning it as the value to the Items property of TodoApp instance.