added TextInput (test and implementation)
Wohoooo, this took most of my Friday. It turns out I had to rethink the TodoItem control, as it has to have one DOM, not several (based on its states). Was I right, or war I wrong? The next steps will tell.

Anyways, TextInput is now integrated into the TodoItem component, great thing.

Oh, and also the Style thing works - had to make a separate xaml file, and use MockApplication to get a cheap Application.Current, and use that xaml file as a Resource Dictionary so that the tests could pick up the style for TextBlock from it.

Not sure if the style is currently done the right way, though... We'll see.